+++ libs/nox-py/src/lib.rs
@@ -15,7 +15,6 @@ use pyo3::prelude::*;
 use pyo3::types::PyAnyMethods;
 
 mod archetype;
-mod asset;
 mod component;
 mod entity;
 mod error;
@@ -27,11 +26,9 @@ mod s10;
 mod spatial;
 mod system;
 mod ukf;
-mod well_known;
 mod world_builder;
 
 pub use archetype::*;
-pub use asset::*;
 pub use component::*;
 pub use entity::*;
 pub use error::*;
@@ -41,7 +38,6 @@ pub use linalg::*;
 pub use query::*;
 pub use spatial::*;
 pub use system::*;
-pub use well_known::*;
 pub use world_builder::*;
 
 trait PyUntypedArrayExt {
@@ -135,21 +131,12 @@ pub fn elodin(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<SpatialMotion>()?;
     m.add_class::<SpatialInertia>()?;
     m.add_class::<Quaternion>()?;
-    m.add_class::<Mesh>()?;
-    m.add_class::<Material>()?;
-    m.add_class::<Handle>()?;
     m.add_class::<PrimitiveType>()?;
     m.add_class::<QueryInner>()?;
     m.add_class::<GraphQueryInner>()?;
     m.add_class::<Edge>()?;
     m.add_class::<Component>()?;
-    m.add_class::<VectorArrow>()?;
-    m.add_class::<BodyAxes>()?;
-    m.add_class::<Color>()?;
-    m.add_class::<Panel>()?;
     m.add_class::<Integrator>()?;
-    m.add_class::<Glb>()?;
-    m.add_class::<Line3d>()?;
     m.add_class::<PyFnSystem>()?;
     m.add_class::<QueryMetadata>()?;
     m.add_class::<SystemBuilder>()?;