+++ libs/nox-py/src/world_builder.rs
@@ -7,7 +7,14 @@ use miette::miette;
 use nox_ecs::{ComponentSchema, IntoSystem, System as _, TimeStep, World, increment_sim_tick, nox};
 use numpy::{PyArray, PyArrayMethods, ndarray::IntoDimension};
 use pyo3::{IntoPyObjectExt, types::PyDict};
-use std::{collections::HashMap, iter, net::SocketAddr, path::PathBuf, time};
+use std::{
+    collections::HashMap,
+    iter,
+    net::SocketAddr,
+    path::{Path, PathBuf},
+    time,
+};
+use tracing::error;
 use zerocopy::{FromBytes, TryFromBytes};
 
 #[derive(Parser, Debug)]
@@ -96,7 +103,6 @@ impl WorldBuilder {
                             component_id,
                             name: component.name.clone(),
                             metadata: component.metadata.clone(),
-                            asset: component.asset,
                         };
 
                         self.world.metadata.component_map.insert(
@@ -117,45 +123,7 @@ impl WorldBuilder {
                 }
                 Ok(())
             }
-            Spawnable::Asset { name, bytes } => {
-                let name = format!("asset_handle_{name}");
-                let component_id = ComponentId::new(&name);
-                let metadata = ComponentMetadata {
-                    component_id,
-                    name,
-                    metadata: Default::default(),
-                    asset: true,
-                };
-
-                self.world.metadata.component_map.insert(
-                    component_id,
-                    (
-                        ComponentSchema {
-                            prim_type: PrimType::U64,
-                            dim: iter::empty().collect(),
-                        },
-                        metadata,
-                    ),
-                );
-                let inner = self.world.assets.insert_bytes(bytes.bytes);
-
-                let buffer = self.world.host.entry(component_id).or_default();
-                buffer.buffer.extend_from_slice(&inner.id.to_le_bytes());
-                buffer
-                    .entity_ids
-                    .extend_from_slice(&entity_id.inner.0.to_le_bytes());
-
-                self.world.dirty_components.insert(component_id);
-
-                Ok(())
-            }
         }
-    }
-
-    fn insert_asset(&mut self, py: Python<'_>, asset: PyObject) -> Result<Handle, Error> {
-        let asset = PyAsset::try_new(py, asset.clone_ref(py))?;
-        let inner = self.world.assets.insert_bytes(asset.bytes()?);
-        Ok(Handle { inner })
     }
 
     fn recipe(&mut self, py: Python<'_>, recipe_obj: PyObject) -> PyResult<()> {
@@ -389,6 +357,33 @@ impl WorldBuilder {
             dict.set_item(comp_name, id.0)?;
         }
         Ok(dict.into_py_any(py)?)
+    }
+
+    /// Set the schematic to represent UI and visualization.
+    ///
+    /// If a `path` is given and the file exists, the file's contents will be
+    /// used as the schematic.
+    ///
+    /// In all other cases, `default_content` is used as the schematic.
+    ///
+    /// Primarily this affords the code a means of specifying a default
+    /// schematic and location for saving custom schematics. It is expected that
+    /// the user may make changes and save the schematic to the given path, but
+    /// this function itself does not write to the `path`.
+    #[pyo3(signature = (default_content = None, path = None,))]
+    pub fn schematic(&mut self, default_content: Option<String>, path: Option<String>) {
+        let file_contents = path.as_deref().and_then(|path| {
+            let path = Path::new(path);
+            if path.exists() {
+                std::fs::read_to_string(path)
+                    .inspect_err(|err| error!(?err, "could not read schematic file at {path:?}"))
+                    .ok()
+            } else {
+                None
+            }
+        });
+        self.world.metadata.schematic_path = path.map(PathBuf::from);
+        self.world.metadata.schematic = file_contents.or(default_content);
     }
 }